Add checks for cases when news has no photos in Edit and Delete method in repository

+++ Infrastructure/JitDalshe.Infrastructure.Persistence/Repositories/NewsRepository.cs
@@ -58,18 +58,29 @@ public sealed class NewsRepository : INewsRepository
 
     public async Task EditAsync(News news, CancellationToken ct = default)
     {
-        var oldPrimaryPhoto = await _dbContext.NewsPrimaryPhotos.FirstAsync(x => x.NewsId == news.Id, ct);
-        _dbContext.NewsPrimaryPhotos.Remove(oldPrimaryPhoto);
+        var oldPrimaryPhoto = await _dbContext.NewsPrimaryPhotos.FirstOrDefaultAsync(x => x.NewsId == news.Id, ct);
+        if (oldPrimaryPhoto is not null)
+        {
+            _dbContext.NewsPrimaryPhotos.Remove(oldPrimaryPhoto);
+        }
 
         _dbContext.News.Update(news);
-        _dbContext.NewsPrimaryPhotos.Add(news.PrimaryPhoto!);
+
+        if (oldPrimaryPhoto is not null)
+        {
+            _dbContext.NewsPrimaryPhotos.Add(news.PrimaryPhoto!);
+        }
 
         await _dbContext.SaveChangesAsync(ct);
     }
 
     public async Task DeleteAsync(News news, CancellationToken ct = default)
     {
-        _dbContext.NewsPrimaryPhotos.Remove(news.PrimaryPhoto!);
+        if (news.PrimaryPhoto is not null)
+        {
+            _dbContext.NewsPrimaryPhotos.Remove(news.PrimaryPhoto);
+        }
+
         _dbContext.NewsPhotos.RemoveRange(news.Photos);
         _dbContext.News.Remove(news);
         await _dbContext.SaveChangesAsync(ct);
